Refactor entities into seperate files
