+++ UI/GlobalUsings.cs
@@ -7,6 +7,7 @@ global using Application.Constants;
 global using Application.Extensions;
 global using Application.Queries;
 global using Application.TuitionPartnerShortlistStorage.Interfaces;
+global using Domain.Constants;
 global using FluentValidation;
 global using FluentValidation.Results;
 global using MediatR;
@@ -15,4 +16,6 @@ global using Microsoft.AspNetCore.Mvc.RazorPages;
 global using Microsoft.EntityFrameworkCore;
 global using UI.Enums;
 global using UI.Extensions;
+global using UI.Mediatr.Queries;
 global using UI.Models;
+global using UI.Structs;
