+++ UI/Structs/LocalAuthorityDistrictCoverage.cs
@@ -1,0 +1,3 @@
+namespace UI.Structs;
+public record struct LocalAuthorityDistrictCoverage
+    (string Region, string Code, string Name, bool InSchool, bool Online);
