+++ UI/UI.csproj
@@ -40,4 +40,8 @@
 		</Content>
 	</ItemGroup>
 
+	<ItemGroup>
+	  <Folder Include="Mediatr\Commands" />
+	</ItemGroup>
+
 </Project>
